+++ ip2geotools/__init__.py
@@ -4,9 +4,9 @@
 __title__ = 'ip2geotools'
 __description__ = 'Simple tool for getting geolocation information on ' + \
                   'given IP address from various geolocation databases.'
-__version__ = '0.1.2'
+__version__ = '0.1.3'
 __author__ = 'Tomas Caha'
 __author_email__ = 'tomas-net@seznam.cz'
 __url__ = 'https://github.com/tomas-net/ip2geotools'
 __license__ = 'MIT License'
-__copyright__ = 'Copyright (c) 2017 Tomas Caha'
+__copyright__ = 'Copyright (c) 2018 Tomas Caha'